requirements.txt
google-generativeai
langchain
chromadb
streamlit
pandas
python-dotenv
tqdm
